Struct tokio_curl::PerformError

pub struct PerformError { /* fields omitted */ }

Error returned by the future returned from perform.

This error can be converted to an io::Error or the underlying Easy handle may also be extracted.

Methods

impl PerformError

fn take_easy(&mut self) -> Option<Easy>

Attempts to extract the underlying Easy handle, if one is available.

For some HTTP requests that fail the Easy handle is still available to recycle for another request. Additionally, the handle may contain information about the failed request. If this is needed, then this method can be called to extract the easy handle.

Note that not all failed HTTP requests will have an easy handle available to return. Some requests may end up destroying the original easy handle as it couldn't be reclaimed.

fn into_error(self) -> Error

Returns the underlying I/O error that caused this error.

All PerformError structures will have an associated I/O error with them. This error indicates why the HTTP request failed, and is likely going to be backed by a curl::Error or a curl::MultiError.

It's also likely if it is available the Easy handle recovered from take_handle will have even more detailed information about the error.